refactor(animation): partition surfaceanimator.cpp behind surfaceanimator_p.h
